Setup init structure to support failable initializers
As of Swift 1.1 (Xcode 6.1) it is not feasible to support failable initializers in combination with non-optional properties. A current limitation is that even failable initializers have to init all non-optional properties first, which makes it impossible to fail an initializer if a non-optional property is not defined in a JSON dictionary. See Xcode 6.1 release notes (https://developer.apple.com/library/ios/releasenotes/DeveloperTools/RN-Xcode/Chapters/Introduction.html#//apple_ref/doc/uid/TP40001051).